Completely refactor the "basic" stepper.
Tests indicated that the "speed" factor for this style of controller cannot be accurately determined across a range of controllers due to timing considerations.
